Introduce trait, move update and create to registry
