+++ pallets/pools-registry/src/lib.rs
@@ -16,7 +16,7 @@ use cfg_primitives::Moment;
 use cfg_traits::Permissions;
 use cfg_types::{PermissionScope, PoolRole, Role};
 use codec::HasCompact;
-use frame_support::{pallet_prelude::*, scale_info::TypeInfo, BoundedVec};
+use frame_support::{pallet_prelude::*, scale_info::TypeInfo, BoundedVec, transactional};
 use frame_system::pallet_prelude::*;
 pub use pallet::*;
 use sp_runtime::{
@@ -52,6 +52,16 @@ where
 	metadata: BoundedVec<u8, MetaSize>,
 }
 
+pub(crate) trait PoolMutate<Balance> {
+	fn create(
+		tranche_inputs: Vec<TranncheInput>,
+		max_reserve: Balance,
+		metadata: Option<Vev<u8>>,
+	) -> DispatchResult;
+
+	fn update() -> DispatchResult;
+}
+
 type PoolMetadataOf<T> = PoolMetadata<<T as Config>::MaxSizeMetadata>;
 
 #[frame_support::pallet]
@@ -81,6 +91,8 @@ pub mod pallet {
 			+ HasCompact
 			+ MaxEncodedLen
 			+ core::fmt::Debug;
+
+		type ModifyPool: PoolMutate<Self::Balance>;
 
 		type CurrencyId: Parameter + Copy;
 
@@ -205,102 +217,7 @@ pub mod pallet {
 				Error::<T>::InvalidCurrency
 			);
 
-			Self::is_valid_tranche_change(
-				None,
-				&tranche_inputs
-					.iter()
-					.map(|t| TrancheUpdate {
-						tranche_type: t.tranche_type,
-						seniority: t.seniority,
-					})
-					.collect(),
-			)?;
-
-			let now = Self::now();
-
-			let tranches = Tranches::from_input::<
-				T::TrancheToken,
-				T::MaxTokenNameLength,
-				T::MaxTokenSymbolLength,
-			>(pool_id, tranche_inputs.clone(), now)?;
-
-			let checked_metadata: Option<BoundedVec<u8, T::MaxSizeMetadata>> = match metadata {
-				Some(metadata_value) => {
-					let checked: BoundedVec<u8, T::MaxSizeMetadata> = metadata_value
-						.try_into()
-						.map_err(|_| Error::<T>::BadMetadata)?;
-
-					Some(checked)
-				}
-				None => None,
-			};
-
-			for (tranche, tranche_input) in tranches.tranches.iter().zip(&tranche_inputs) {
-				let token_name: BoundedVec<u8, T::MaxTokenNameLength> =
-					tranche_input.clone().metadata.token_name.clone();
-
-				let token_symbol: BoundedVec<u8, T::MaxTokenSymbolLength> =
-					tranche_input.metadata.token_symbol.clone();
-
-				let decimals = match T::AssetRegistry::metadata(&currency) {
-					Some(metadata) => metadata.decimals,
-					None => return Err(Error::<T>::MetadataForCurrencyNotFound.into()),
-				};
-
-				let parachain_id = T::ParachainId::get();
-
-				let metadata = tranche.create_asset_metadata(
-					decimals,
-					parachain_id,
-					token_name.to_vec(),
-					token_symbol.to_vec(),
-				);
-
-				T::AssetRegistry::register_asset(Some(tranche.currency), metadata)
-					.map_err(|_| Error::<T>::FailedToRegisterTrancheMetadata)?;
-			}
-
-			Pool::<T>::insert(
-				pool_id,
-				PoolDetails {
-					currency,
-					tranches,
-					status: PoolStatus::Open,
-					epoch: EpochState {
-						current: One::one(),
-						last_closed: now,
-						last_executed: Zero::zero(),
-					},
-					parameters: PoolParameters {
-						min_epoch_time: sp_std::cmp::min(
-							sp_std::cmp::max(
-								T::DefaultMinEpochTime::get(),
-								T::MinEpochTimeLowerBound::get(),
-							),
-							T::MinEpochTimeUpperBound::get(),
-						),
-						max_nav_age: sp_std::cmp::min(
-							T::DefaultMaxNAVAge::get(),
-							T::MaxNAVAgeUpperBound::get(),
-						),
-					},
-					reserve: ReserveDetails {
-						max: max_reserve,
-						available: Zero::zero(),
-						total: Zero::zero(),
-					},
-					metadata: checked_metadata,
-				},
-			);
-
-			T::Permission::add(
-				PermissionScope::Pool(pool_id),
-				admin.clone(),
-				Role::PoolRole(PoolRole::PoolAdmin),
-			)?;
-
-			Self::deposit_event(Event::Created { pool_id, admin });
-			Ok(())
+			T::ModifyPool::create(tranche_inputs, max_reserve, metadata)
 		}
 
 		/// Update per-pool configuration settings.
