+++ pallets/pools/Cargo.toml
@@ -12,31 +12,31 @@ version = '3.0.0'
 targets = ['x86_64-unknown-linux-gnu']
 
 [dependencies]
-serde = { version = "1.0.119" }
+cfg-primitives = { path = "../../libs/primitives", default-features = false}
+cfg-traits = { path = "../../libs/traits", default-features = false }
+cfg-types = { path = "../../libs/types", default-features = false }
 codec = { package = 'parity-scale-codec', version = '3.0.0', features = ['derive'] , default-features = false }
-scale-info = { version = "2.0", default-features = false, features = ["derive"] }
-rev_slice = { version = "0.1.5", default-features = false }
-lazy_static = { version = "1.4.0", optional = true, default-features = false }
-
 frame-support = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
 frame-system = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
+lazy_static = { version = "1.4.0", optional = true, default-features = false }
+orml-asset-registry = {  git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.29" }
+orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.29" }
+pallet-permissions = { path = "../../pallets/permissions", default-features = false}
+pallet-pools-registry = { path = "../../pallets/pools-registry", default-features = false }
 pallet-timestamp = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
+polkadot-parachain = { git = "https://github.com/paritytech/polkadot", default-features = false , branch = "release-v0.9.29" }
+rev_slice = { version = "0.1.5", default-features = false }
+scale-info = { version = "2.0", default-features = false, features = ["derive"] }
+serde = { version = "1.0.119" }
 sp-arithmetic = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
 sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
 sp-std = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
-orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.29" }
-cfg-traits = { path = "../../libs/traits", default-features = false }
-cfg-primitives = { path = "../../libs/primitives", default-features = false}
-cfg-types = { path = "../../libs/types", default-features = false }
-
-pallet-permissions = { path = "../../pallets/permissions", default-features = false}
 xcm = { git = "https://github.com/paritytech/polkadot", default-features = false , branch = "release-v0.9.29" }
-polkadot-parachain = { git = "https://github.com/paritytech/polkadot", default-features = false , branch = "release-v0.9.29" }
-orml-asset-registry = {  git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.29" }
 
 # Benchmarking dependencies - optional
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", default-features = false , optional = true , branch = "polkadot-v0.9.29" }
 pallet-investments = { path = "../../pallets/investments", default-features = false, optional = true}
+
 
 [dev-dependencies]
 cfg-test-utils = { path = "../../libs/test-utils",default-features = true }
